+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/config/CustomUserDetailsService.java
@@ -1,9 +1,11 @@
-package com.crashcourse.kickoff.tms.user;
+package com.crashcourse.kickoff.tms.user.config;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import com.crashcourse.kickoff.tms.user.repository.UserRepository;
 
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
