+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/dto/NewUserDTO.java
@@ -1,4 +1,4 @@
-package com.crashcourse.kickoff.tms.user;
+package com.crashcourse.kickoff.tms.user.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,7 +6,7 @@ import lombok.Setter;
 // Data Transfer Object to facilitate setting user roles on server side (allows for HTTP post requests with only username and password)
 @Setter
 @Getter
-public class UserDTO {
+public class NewUserDTO {
     private String username;
     private String password;
 }
