+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/model/Role.java
@@ -1,4 +1,4 @@
-package com.crashcourse.kickoff.tms.user;
+package com.crashcourse.kickoff.tms.user.model;
 
 public enum Role {
     ROLE_ADMIN,
