+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/repository/UserRepository.java
@@ -1,9 +1,11 @@
-package com.crashcourse.kickoff.tms.user;
+package com.crashcourse.kickoff.tms.user.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.crashcourse.kickoff.tms.user.model.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
