+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/service/UserService.java
@@ -1,0 +1,12 @@
+package com.crashcourse.kickoff.tms.user.service;
+
+import java.util.List;
+
+import com.crashcourse.kickoff.tms.user.dto.NewUserDTO;
+import com.crashcourse.kickoff.tms.user.model.User;
+
+// Business logic of UserService, actual implementation in UserServiceImpl
+public interface UserService {
+    public List<User> getUsers();
+    User addUser(NewUserDTO newUserDTO);
+}
